requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.20
python-dotenv==1.0.0
Werkzeug==2.3.7
PyJWT==2.8.0
reportlab==4.0.4
python-dateutil==2.8.2
Babel==2.13.0  # For currency formatting